move module to utilities, add docstrings to tops of mods
